1.1.0, Major changes
Finally, a working spambot. Comfort features are still remaining but the
basic functionality is completely done.

+++ src/index.ts
@@ -1,4 +1,4 @@
-console.log("INFO: Starting...");
+console.log("INFO: Starting.");
 
 import { Client, DMChannel, GroupDMChannel, TextChannel, User } from 'discord.js';
 import { start, token } from '../config/config.json';
@@ -11,10 +11,13 @@ let store: {
 };
 const client = new Client();
 
-client.login(token)
-    .then(() => console.log("SUCCESS: Succesfully logged in."))
-    .catch((res) => console.log(`ERROR: Couldn't log in, check your credentials.\
+if (!args.debug)
+    client.login(token)
+        .then(() => console.log("SUCCESS: Succesfully logged in."))
+        .catch((res) => console.log(`ERROR: Couldn't log in, check your credentials.\
 ${args.verbose ? `\n${res}` : ''}`));
+
+else console.log("INFO: Skipping login as debug is specified.");
 
 const messenger = new Messenger( //@ts-ignore DUDE I AM CHECKING IF ITS NULL IN ARGS.TS BUDDY HELLO TYPESCRIPT BRUH?
     args.list ? args.list : args.msg,
@@ -25,27 +28,32 @@ const messenger = new Messenger( //@ts-ignore DUDE I AM CHECKING IF ITS NULL IN 
 );
 
 messenger.on('send', msg => {
-    store.channel.send(msg)
-        .then(() => {
+    if (!args.debug)
+        store.channel.send(msg)
+            .then(() => {
+                if (args.verbose)
+                    console.log("SUCCESSINFO: Sent message " + msg + '.');
+            });
+    else console.log("INFO: Send event emitted " + msg + ".")
+});
+
+if (!args.debug)
+    client.on('message', (message) => {
+        if (message.content == start) {
             if (args.verbose)
-                console.log("SUCCESSINFO: Sent message " + msg + '.');
-        });
-});
+                console.log(`INFO: Received command at ${Date()} in ${message.channel.type == "text" //@ts-ignore	
+                    ? `channel ${message.channel.name}.` //I think TS doesn't know about if statements
+                    : `a DM channel.`
+                    }`);
 
-client.on('message', (message) => {
-    if (message.content == start) {
-        if (args.verbose)
-            console.log(`INFO: Received command at ${Date()} in ${message.channel.type == "text" //@ts-ignore	
-                ? `channel ${message.channel.name}.` //I think TS doesn't know about if statements
-                : `a DM channel.`
-                }`);
+            store = {
+                channel: message.channel,
+                author: message.author
+            };
 
-        message.delete();
-        messenger.start();
-
-        store = {
-            channel: message.channel,
-            author: message.author
+            message.delete();
+            messenger.start();
         }
-    }
-});
+    });
+
+else messenger.start();
